Add files via upload
When the game starts, it asks you to enter a money amount. Then 3 random fruits will appear on the screen. If two of these fruits are the same, the money is doubled. If all three are the same, the money is tripled. If none of them are the same, the money is lost.